requirements.txt
pillow>=9.0.0
tkinterdnd2>=0.3.0
